Enum classes patch for Nintendo Switch port

+++ GsKit/base/CInput.cpp
@@ -81,42 +81,42 @@ void CInput::resetControls(const int player)
 
 #ifdef __SWITCH__
 	// Switch gamepad mapping using two joycons
-    curInput[IC_LEFT].joyeventtype = EType::JOYBUTTON;
+    curInput[IC_LEFT].joyeventtype = (unsigned int)EType::JOYBUTTON;
 	curInput[IC_LEFT].joybutton = 12;
 	curInput[IC_LEFT].which = 0;
-    curInput[IC_UP].joyeventtype = EType::JOYBUTTON;
+    curInput[IC_UP].joyeventtype = (unsigned int)EType::JOYBUTTON;
 	curInput[IC_UP].joybutton = 13;
 	curInput[IC_UP].which = 0;
-    curInput[IC_RIGHT].joyeventtype = EType::JOYBUTTON;
+    curInput[IC_RIGHT].joyeventtype = (unsigned int)EType::JOYBUTTON;
 	curInput[IC_RIGHT].joybutton = 14;
 	curInput[IC_RIGHT].which = 0;
-    curInput[IC_DOWN].joyeventtype = EType::JOYBUTTON;
+    curInput[IC_DOWN].joyeventtype = (unsigned int)EType::JOYBUTTON;
 	curInput[IC_DOWN].joybutton = 15;
 	curInput[IC_DOWN].which = 0;
 
-    curInput[IC_JUMP].joyeventtype = EType::JOYBUTTON;
+    curInput[IC_JUMP].joyeventtype = (unsigned int)EType::JOYBUTTON;
 	curInput[IC_JUMP].joybutton = 0;
 	curInput[IC_JUMP].which = 0;
-    curInput[IC_POGO].joyeventtype = EType::JOYBUTTON;
+    curInput[IC_POGO].joyeventtype = (unsigned int)EType::JOYBUTTON;
 	curInput[IC_POGO].joybutton = 2;
 	curInput[IC_POGO].which = 0;
-    curInput[IC_FIRE].joyeventtype = EType::JOYBUTTON;
+    curInput[IC_FIRE].joyeventtype = (unsigned int)EType::JOYBUTTON;
 	curInput[IC_FIRE].joybutton = 3;
 	curInput[IC_FIRE].which = 0;
-    curInput[IC_RUN].joyeventtype = EType::JOYBUTTON;
+    curInput[IC_RUN].joyeventtype = (unsigned int)EType::JOYBUTTON;
     curInput[IC_RUN].joybutton = 1;
 	curInput[IC_RUN].which = 0;
 
-    curInput[IC_STATUS].joyeventtype = EType::JOYBUTTON;
+    curInput[IC_STATUS].joyeventtype = (unsigned int)EType::JOYBUTTON;
 	curInput[IC_STATUS].joybutton = 11;
 	curInput[IC_STATUS].which = 0;
-    curInput[IC_CAMLEAD].joyeventtype = EType::JOYBUTTON;
+    curInput[IC_CAMLEAD].joyeventtype = (unsigned int)EType::JOYBUTTON;
 	curInput[IC_CAMLEAD].joybutton = 7;
 	curInput[IC_CAMLEAD].which = 0;
-    curInput[IC_HELP].joyeventtype = EType::JOYBUTTON;
+    curInput[IC_HELP].joyeventtype = (unsigned int)EType::JOYBUTTON;
 	curInput[IC_HELP].joybutton = 6;
 	curInput[IC_HELP].which = 0;
-    curInput[IC_BACK].joyeventtype = EType::JOYBUTTON;
+    curInput[IC_BACK].joyeventtype = (unsigned int)EType::JOYBUTTON;
 	curInput[IC_BACK].joybutton = 10;
 	curInput[IC_BACK].which = 0;
 #else
